// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetMandatoryComponentsListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { MandatoryComponentsList } from './MandatoryComponentsList';
import {
    MandatoryComponentsListFromJSON,
    MandatoryComponentsListFromJSONTyped,
    MandatoryComponentsListToJSON,
} from './MandatoryComponentsList';

/**
 * 
 * @export
 * @interface GetMandatoryComponentsListResponse
 */
export interface GetMandatoryComponentsListResponse {
    /**
     * 
     * @type {MandatoryComponentsList}
     * @memberof GetMandatoryComponentsListResponse
     */
    mandatoryComponentsList: MandatoryComponentsList;
}

/**
 * Check if a given object implements the GetMandatoryComponentsListResponse interface.
 */
export function instanceOfGetMandatoryComponentsListResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "mandatoryComponentsList" in value;

    return isInstance;
}

export function GetMandatoryComponentsListResponseFromJSON(json: any): GetMandatoryComponentsListResponse {
    return GetMandatoryComponentsListResponseFromJSONTyped(json, false);
}

export function GetMandatoryComponentsListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetMandatoryComponentsListResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'mandatoryComponentsList': MandatoryComponentsListFromJSON(json['mandatoryComponentsList']),
    };
}

export function GetMandatoryComponentsListResponseToJSON(value?: GetMandatoryComponentsListResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'mandatoryComponentsList': MandatoryComponentsListToJSON(value.mandatoryComponentsList),
    };
}