// frontend/web/src/services/API/proserve-wb-packaging-api/models/UpdateMandatoryComponentsListRequest.ts
export interface UpdateMandatoryComponentsListRequest {
    /**
     * Platform for which the mandatory components list is enforced
     * @type {string}
     * @memberof UpdateMandatoryComponentsListRequest
     */
    mandatoryComponentsListPlatform: string;
    /**
     * OS for which the mandatory components list is enforced
     * @type {string}
     * @memberof UpdateMandatoryComponentsListRequest
     */
    mandatoryComponentsListOsVersion: string;
    /**
     * Architecture for which the mandatory components list is enforced
     * @type {string}
     * @memberof UpdateMandatoryComponentsListRequest
     */
    mandatoryComponentsListArchitecture: string;
    /**
     * Components versions that will be prepended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof UpdateMandatoryComponentsListRequest
     */
    prependedComponentsVersions: Array<ComponentVersionEntry>;
    /**
     * Components versions that will be appended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof UpdateMandatoryComponentsListRequest
     */
    appendedComponentsVersions: Array<ComponentVersionEntry>;
}